+++ Global/Charts/2013/BarChart.cs
@@ -8,7 +8,7 @@ namespace OpenXMLOffice.Global_2013
 	/// <summary>
 	/// Represents the settings for a bar chart.
 	/// </summary>
-	public class BarFamilyChart : ChartBase
+	public class BarChart : ChartBase
 	{
 		private const int DefaultGapWidth = 150;
 		private const int DefaultOverlap = 100;
@@ -18,7 +18,7 @@ namespace OpenXMLOffice.Global_2013
 		/// </summary>
 		protected readonly BarChartSetting barChartSetting;
 
-		internal BarFamilyChart(BarChartSetting barChartSetting) : base(barChartSetting)
+		internal BarChart(BarChartSetting barChartSetting) : base(barChartSetting)
 		{
 			this.barChartSetting = barChartSetting;
 		}
@@ -26,7 +26,7 @@ namespace OpenXMLOffice.Global_2013
 		/// <summary>
 		/// Create Bar Chart with provided settings
 		/// </summary>
-		protected BarFamilyChart(BarChartSetting barChartSetting, ChartData[][] dataCols) : base(barChartSetting)
+		public BarChart(BarChartSetting barChartSetting, ChartData[][] dataCols) : base(barChartSetting)
 		{
 			this.barChartSetting = barChartSetting;
 			SetChartPlotArea(CreateChartPlotArea(dataCols));
